Add slice (for general error) to ssz util
