+++ beacon-chain/p2p/types/BUILD.bazel
@@ -29,6 +29,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//shared/params:go_default_library",
+        "//shared/sszutil:go_default_library",
         "//shared/testutil/assert:go_default_library",
         "//shared/testutil/require:go_default_library",
     ],
