+++ beacon-chain/p2p/types/types_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/prysmaticlabs/prysm/shared/params"
+	"github.com/prysmaticlabs/prysm/shared/sszutil"
 	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
 )
@@ -24,17 +25,6 @@ func TestBeaconBlockByRootsReq_Limit(t *testing.T) {
 	}
 	req2 := BeaconBlockByRootsReq(nil)
 	require.ErrorContains(t, "expected buffer with length of upto", req2.UnmarshalSSZ(buf))
-}
-
-func TestErrorResponse_Limit(t *testing.T) {
-	errorMessage := make([]byte, 0)
-	// Provide a message of size 6400 bytes.
-	for i := uint64(0); i < 200; i++ {
-		byteArr := [32]byte{byte(i)}
-		errorMessage = append(errorMessage, byteArr[:]...)
-	}
-	errMsg := ErrorMessage{}
-	require.ErrorContains(t, "expected buffer with length of upto", errMsg.UnmarshalSSZ(errorMessage))
 }
 
 func TestRoundTripSerialization(t *testing.T) {
@@ -59,12 +49,12 @@ func roundTripTestBlocksByRootReq(t *testing.T) {
 
 func roundTripTestErrorMessage(t *testing.T) {
 	errMsg := []byte{'e', 'r', 'r', 'o', 'r'}
-	sszErr := make(ErrorMessage, len(errMsg))
+	sszErr := make(sszutil.ErrorMessage, len(errMsg))
 	copy(sszErr, errMsg)
 
 	marshalledObj, err := sszErr.MarshalSSZ()
 	require.NoError(t, err)
-	newVal := ErrorMessage(nil)
+	newVal := sszutil.ErrorMessage(nil)
 
 	require.NoError(t, newVal.UnmarshalSSZ(marshalledObj))
 	assert.DeepEqual(t, []byte(newVal), errMsg)
