+++ beacon-chain/sync/error_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	p2ptest "github.com/prysmaticlabs/prysm/beacon-chain/p2p/testing"
-	"github.com/prysmaticlabs/prysm/beacon-chain/p2p/types"
+	"github.com/prysmaticlabs/prysm/shared/sszutil"
 	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
 )
@@ -22,7 +22,7 @@ func TestRegularSync_generateErrorResponse(t *testing.T) {
 	_, err = buf.Read(b)
 	require.NoError(t, err)
 	assert.Equal(t, responseCodeServerError, b[0], "The first byte was not the status code")
-	msg := &types.ErrorMessage{}
+	msg := &sszutil.ErrorMessage{}
 	require.NoError(t, r.p2p.Encoding().DecodeWithMaxLength(buf, msg))
 	assert.Equal(t, "something bad happened", string(*msg), "Received the wrong message")
 }
